refactor: moved logger into a more sensible `lib` directory
