+++ src/main.ts
@@ -1,7 +1,7 @@
 import * as CreepManager from "./components/creeps/creepManager";
 import * as Config from "./config/config";
 
-import { log } from "./components/support/log";
+import { log } from "./lib/logger/log";
 
 // Any code written outside the `loop()` method is executed only when the
 // Screeps system reloads your script.
